+++ CHANGELOG.md
@@ -1,12 +1,18 @@
 # SelfieStop Mobile App
 
+## v0.1.1
+
+- Added Splash Screen with logo and loading indicator
+- Added Login Tab
+- Added Register Tab
+
 ## v0.1.0
 
-- Removed non-functioning buttons, added navigation upon callout press
+- Removed non-functioning buttons, added navigation upon call out press
 
 ## v0.0.8
 
-- Added styles, navbar, titles, full add stop functionality, and new map with custom callouts containing images.
+- Added styles, navbar, titles, full add stop functionality, and new map with custom call outs containing images.
 
 ## v0.0.5